Add one folder for numerical method (prevision for DG)
