+++ src/coreComponents/physicsSolvers/wavePropagation/sem/acoustic/secondOrderEqn/vti/AcousticVTIWaveEquationSEM.cpp
@@ -27,7 +27,7 @@
 #include "mesh/ElementType.hpp"
 #include "mesh/mpiCommunications/CommunicationTools.hpp"
 #include "physicsSolvers/wavePropagation/shared/WaveSolverUtils.hpp"
-#include "physicsSolvers/wavePropagation/acoustic/shared/AcousticTimeSchemeSEMKernel.hpp"
+#include "physicsSolvers/wavePropagation/sem/acoustic/shared/AcousticTimeSchemeSEMKernel.hpp"
 
 namespace geos
 {
