+++ src/coreComponents/physicsSolvers/wavePropagation/sem/acoustic/secondOrderEqn/vti/AcousticVTIWaveEquationSEM.hpp
@@ -23,7 +23,7 @@
 #include "mesh/MeshFields.hpp"
 #include "physicsSolvers/SolverBase.hpp"
 #include "physicsSolvers/wavePropagation/shared/WaveSolverBase.hpp"
-#include "physicsSolvers/wavePropagation/acoustic/shared/AcousticFields.hpp"
+#include "physicsSolvers/wavePropagation/sem/acoustic/shared/AcousticFields.hpp"
 #include "AcousticVTIFields.hpp"
 
 namespace geos
